updated main

+++ ft_printf.h
@@ -8,7 +8,7 @@ int ft_putchar(char c);
 size_t ft_printstr(char *str);
 size_t ft_printnbr(int nbr);
 size_t ft_print_unsigned(unsigned int nbr);
-size_t ft_print_hex(int long num, int format);
+size_t ft_print_hex(unsigned int num, int format);
 size_t ft_count_words(char * str);
 
 #endif
